Implement a strategy to remove old bytes on the serial bus
Problem: when the arduino restarts, the end of that last communication is still buffered (arduino side ? Windows side ?)
Solution(?): when the COM port is opened (windows-side), I trash all bytes until a silent period is observed. (a few ms).
The arduino must not start any useful communication before that.
In practice, the setup phase is quite long, orders of magnitude longer than the required silent period. The 'delay' on arduino is probably 100% useless.